+++ templates/web-angular/web/main.dart
@@ -1,8 +1,6 @@
 import 'package:angular/angular.dart';
-import 'package:__projectName__/app_component.dart';
-
-import 'main.template.dart' as ng;
+import 'package:__projectName__/app_component.template.dart' as ng;
 
 void main() {
-  bootstrapStatic(AppComponent, [], ng.initReflector);
+  runApp(ng.AppComponentNgFactory);
 }